Align Open API specs, Entity and DTO numeric types with DB sizes to avoid excessive conversions.

+++ src/main/java/uk/gov/hmcts/opal/dto/MinorCreditorSearch.java
@@ -2,12 +2,11 @@ package uk.gov.hmcts.opal.dto;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
-import java.util.List;
 
 @Data
 @NoArgsConstructor
@@ -17,7 +16,7 @@ import java.util.List;
 public class MinorCreditorSearch implements ToJsonString {
 
     @JsonProperty("business_unit_ids")
-    private List<Integer> businessUnitIds;
+    private List<Short> businessUnitIds;
 
     @JsonProperty("active_accounts_only")
     private Boolean activeAccountsOnly;
