+++ src/test/java/uk/gov/hmcts/opal/service/DefendantAccountServiceTest.java
@@ -267,7 +267,7 @@ class DefendantAccountServiceTest {
         PostDefendantAccountSearchRequestDefendantAccount request =
             PostDefendantAccountSearchRequestDefendantAccount.builder()
                 .activeAccountsOnly(true)
-                .businessUnitIds(List.of(77))
+                .businessUnitIds(List.of((short) 77))
                 .referenceNumber(new DefendantAccountSearchReferenceNumberDefendantAccount()
                     .organisation(false)
                     .accountNumber("A123"))
